+++ cli/commands/check/help.go
@@ -3,6 +3,7 @@ package check
 import (
 	"github.com/sensu/sensu-go/cli"
 	"github.com/sensu/sensu-go/cli/commands/check/subcommands"
+	"github.com/sensu/sensu-go/cli/commands/helpers"
 	"github.com/spf13/cobra"
 )
 
@@ -11,6 +12,7 @@ func HelpCommand(cli *cli.SensuCli) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "check",
 		Short: "Manage checks",
+		RunE:  helpers.DefaultSubCommandRunE,
 	}
 
 	// Add sub-commands