+++ cli/commands/root/root.go
@@ -18,6 +18,12 @@ func Command() *cobra.Command {
 		SilenceUsage: true,
 		Run: func(cmd *cobra.Command, args []string) {
 			_ = cmd.Help()
+			// JK: Ideally we could return command.UsageError here but we cannot
+			// silence errors in the root command like we do with subcommands.
+			// When SilenceErrors is set to true in the root command it will
+			// set SilenceErrors = true for all subcommands. As a result, this
+			// is the only command containing subcommands that will exit 0 when
+			// no arguments are given.
 		},
 	}